cleanup: remove unnecessary protoxform
