+++ tools/protoprint/utils.py
