+++ tools/testdata/protoxform/envoy/v2/discovery_service.proto.gold
