+++ tools/testdata/protoxform/envoy/v2/fully_qualified_names.proto.gold
