+++ tools/testdata/protoxform/envoy/v2/oneof.proto.gold
