+++ tools/testdata/protoxform/envoy/v2/package_move.proto.gold
