+++ tools/testdata/protoxform/envoy/v2/sample.proto.gold
